+++ lib/app_storage_kit.dart
@@ -7,3 +7,5 @@ export 'package:app_storage_kit/db_storage.dart';
 export 'package:app_storage_kit/data_models/db_record.dart';
 export 'package:app_storage_kit/data_models/db_table_model.dart';
 export 'package:app_storage_kit/data_models/table_field_model.dart';
+
+export 'package:sqflite_common/src/sql_builder.dart';